# Avoid unnecessary sleep when handling caret movement to improve responsiveness.