--- a/source/editableText.py
+++ b/source/editableText.py
@@ -47,10 +47,10 @@ class EditableText(TextContainerObject,ScriptableObject):
 	announceEntireNewLine=False
 
 	#: The minimum amount of time that should elapse before checking if the word under the caret has changed
-	_hasCaretMoved_minWordTimeoutMs=30
+	_hasCaretMoved_minWordTimeoutSec = 0.03
 
 	#: The maximum amount of time that may elapse before we no longer rely on caret events to detect movement.
-	_useEvents_maxTimeoutMs = 10
+	_useEvents_maxTimeoutSec = 0.06
 
 	_caretMovementTimeoutMultiplier = 1
 
@@ -71,22 +71,18 @@ class EditableText(TextContainerObject,ScriptableObject):
 		@rtype: tuple
 		"""
 		if timeout is None:
-			timeoutMs = config.conf["editableText"]["caretMoveTimeoutMs"]
-		else:
-			# This function's arguments are in seconds, but we want ms.
-			timeoutMs = timeout * 1000
-		timeoutMs *= self._caretMovementTimeoutMultiplier
-		# time.sleep accepts seconds, so retryInterval is in seconds.
-		# Convert to integer ms to avoid floating point precision errors when adding to elapsed.
-		retryMs = int(retryInterval * 1000)
+			timeout = config.conf["editableText"]["caretMoveTimeoutMs"] / 1000
+		timeout *= self._caretMovementTimeoutMultiplier
+		start = time.time()
 		elapsed = 0
 		newInfo=None
+		retries = 0
 		while True:
 			if isScriptWaiting():
 				return (False,None)
 			api.processPendingEvents(processEventQueue=False)
 			if eventHandler.isPendingEvents("gainFocus"):
-				log.debug("Focus event. Elapsed: %d ms" % elapsed)
+				log.debug("Focus event. Elapsed %g sec" % elapsed)
 				return (True,None)
 			# If the focus changes after this point, fetching the caret may fail,
 			# but we still want to stay in this loop.
@@ -97,14 +93,17 @@ class EditableText(TextContainerObject,ScriptableObject):
 			else:
 				# Caret events are unreliable in some controls.
 				# Only use them if we consider them safe to rely on for a particular control,
-				# and only if they arrive within C{_useEvents_maxTimeoutMs} mili seconds
+				# and only if they arrive within C{_useEvents_maxTimeoutSec} seconds
 				# after causing the event to occur.
 				if (
-					elapsed <= self._useEvents_maxTimeoutMs and
-					self.caretMovementDetectionUsesEvents and
-					(eventHandler.isPendingEvents("caret") or eventHandler.isPendingEvents("textChange"))
+					elapsed <= self._useEvents_maxTimeoutSec
+					and self.caretMovementDetectionUsesEvents
+					and (eventHandler.isPendingEvents("caret") or eventHandler.isPendingEvents("textChange"))
 				):
-					log.debug("Caret move detected using event. Elapsed: %d ms" % elapsed)
+					log.debug(
+						"Caret move detected using event. Elapsed %g sec, retries %d"
+						% (elapsed, retries)
+					)
 					return (True,newInfo)
 			# Try to detect with bookmarks.
 			newBookmark = None
@@ -114,9 +113,12 @@ class EditableText(TextContainerObject,ScriptableObject):
 				except (RuntimeError,NotImplementedError):
 					pass
 			if newBookmark and newBookmark!=bookmark:
-				log.debug("Caret move detected using bookmarks. Elapsed: %d ms" % elapsed)
+				log.debug(
+					"Caret move detected using bookmarks. Elapsed %g sec, retries %d"
+					% (elapsed, retries)
+				)
 				return (True, newInfo)
-			if origWord is not None and newInfo and elapsed >= self._hasCaretMoved_minWordTimeoutMs:
+			if origWord is not None and newInfo and elapsed >= self._hasCaretMoved_minWordTimeoutSec:
 				# When pressing delete, bookmarks might not be enough to detect caret movement.
 				# Therefore try detecting if the word under the caret has changed, such as when pressing delete.
 				# some editors such as Mozilla Gecko can have text and units that get out of sync with eachother while a character is being deleted.
@@ -125,13 +127,21 @@ class EditableText(TextContainerObject,ScriptableObject):
 				wordInfo.expand(textInfos.UNIT_WORD)
 				word = wordInfo.text
 				if word != origWord:
-					log.debug("Word at caret changed. Elapsed: %d ms" % elapsed)
+					log.debug("Word at caret changed. Elapsed: %g sec" % elapsed)
 					return (True, newInfo)
-			if elapsed >= timeoutMs:
+			elapsed = time.time() - start
+			if elapsed >= timeout:
 				break
-			time.sleep(retryInterval)
-			elapsed += retryMs
-		log.debug("Caret didn't move before timeout. Elapsed: %d ms" % elapsed)
+			# We spin the first few tries, as sleep is not accurate for tiny periods
+			# and we might end up sleeping longer than we need to. Spinning improves
+			# responsiveness in the case that the app responds fairly quickly.
+			if retries > 2:
+				# Don't spin too long, though. If we get to this point, the app is
+				# probably taking a while to respond, so super fast response is
+				# already lost.
+				time.sleep(retryInterval)
+			retries += 1
+		log.debug("Caret didn't move before timeout. Elapsed: %g sec" % elapsed)
 		return (False,newInfo)
 
 	def _caretScriptPostMovedHelper(self, speakUnit, gesture, info=None):