@jcsteh 2023-03-09
I have tested quite a bit with holding down the downArrow key. I haven't tested with control+downArrow though. I'll do that. It sounds like I should also test with control+shift+downArrow based on #11478.