@jcsteh 2023-03-09
I've just held down control+downArrow and control+shift+downArrow for a long time in a pretty long document and didn't encounter any freeze or anything like that. Still, I'll watch out for it while using it in daily work.