@jcsteh 2023-03-09
I doubt it. This PR is about improving the response time between pressing a key and NVDA checking whether the cursor has moved. That issue sounds like it's more about NVDA not being able to get correct info from the editor for somereason. If anything, that issue probably requires a slower response time.