@amirmahdifard 2023-03-09
thank you for this, i also have some delay with pressing keys and getting responce from nvda sometimes, i hope this will fix that problem: thanks?