@codeofdusk 2023-03-11
This PR makes tracking of deleted characters in UIA Word much less reliable:

1. Open a new email message in Outlook.
2. Tab to the body.
3. Enter "abc", then press backspace. Observe that the deletion of anything from zero to all three characters are reported.
4. Repeat steps 2 and 3. Observe possibly different results.