@jcsteh 2023-03-11
Is this with the latest commit in this branch? If so, we might need to raise `_useEvents_maxTimeoutSec` even further to 0.03. I already raised it to 0.02, but it might not be enough. I explained why this is necessary in the edited PR description:

> 4. Increase the use events timeout slightly from 10 ms to 20 ms. As noted in 3), we were previously multiplying the sleep amount, which meant our supposed elapsed time was actually less than what had truly elapsed. This just updates the timeout to reflect reality. Without this, backspacing occasionally failed in UIA Notepad.