@jcsteh 2023-03-11
To explain this further, we previously had a retry interval of 0.01 and we just kept multiplying that by the number of retries to calculate the elapsed time; e.g. 0.01, 0.02, 0.03, etc. The problem is that because of the Windows timer resolution, this could be quite inaccurate. If each sleep took 16 ms instead of 10 ms, the elapsed time for 3 tries would be 0.048, 18 ms longer than we calculated.