@codeofdusk 2023-03-13
> Is this with the latest commit in this branch?

Yes it was. Initial testing with 0.03 looks promising, but I'll run with it for a while to be sure.