@codeofdusk 2023-03-17
@jcsteh Found aanother bug with the latest commit in this branch:

1. Open a folder in Windows Explorer.
2. Tab to the search field.
3. enter "abc".
4. Press backspace three times to delete all three characters.

Expected: NVDA reports "c", "b", "a".

Actual: NVDA reports "c", "b".