@jcsteh 2023-03-18
I couldn't reproduce the search bug exactly as you described. However, I did see intermittent failures when backspacing.

To address this, I've had to raise the event timeout even higher to 60 ms. It looks like that control responds very slowly, both in firing events and in processing queries. Not only did our elapsed time previously not account for sleep inaccuracy, it also didn't account for query time.

I'm actually not quite sure why we have that event timeout at all. I looked at the various pull requests which introduced the caret event changes for UIA, but I don't see a justification for the event timeout.