@michaelDCurran 2023-03-23
@codeofdusk  is the explorer search field working okay for you now with the latest fix from @jcsteh  ?